requirements.txt
streamlit
python-dotenv
google-generativeai
PyPDF2
